# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ProxyLoom P0 API
  version: 1.0.0
  description: >-
    P0 contract for a single configured workspace and trusted administrators.
    Five administrator authentication operations are implemented. Other operations
    remain contract-only; declarations do not assert kernel capability or verification.
    Management uses the session Cookie; internal operations exist exclusively on
    a separate mTLS listener; subscription retrieval uses its mandatory path token
    independently of either credential. Scope and resource IDs are server assigned.
    JSON decoding rejects unknown fields, duplicate keys, trailing documents,
    invalid Unicode, and undocumented nulls. Errors never echo input values.
    All references in this document are local; schema validation must not fetch URLs.
servers:
  - url: /
    description: Management and public subscription listener; internal paths use their own listener.
security:
  - SessionCookie: []
...
components:
  securitySchemes:
    SessionCookie:
      type: "apiKey"
      in: "cookie"
      name: "proxyloom_session"
      description: "Opaque server-side administrator session; Secure, HttpOnly, SameSite=Lax. State-changing management operations additionally enforce Origin and X-CSRF-Token. Setup/login require Origin and JSON but no session CSRF token. Secure may be omitted only for explicit development HTTP loopback PUBLIC_URL. Never accepted as subscription or Runner authorization."
    RunnerMTLS:
      type: "mutualTLS"
      description: "Registered client certificate verified only on the dedicated internal listener. Certificate identity and server registration determine eligible builds, task types and slots. No management Cookie or subscription token is accepted."
  headers:
    RequestID:
      description: "Server-generated safe request correlation ID."
      schema: { $ref: "#/components/schemas/RequestID" }
...
    NoStore:
      schema:
        type: "string"
        const: "private, no-store"
    RetryAfter:
      description: "Bounded retry delay in seconds."
      schema:
        type: "integer"
        minimum: 1
        maximum: 86400
...
  responses:
    BadRequest:
      description: "Malformed JSON, duplicate/unknown field, unsafe shape, invalid path/query/header, invalid Unicode or cursor. No submitted key or value is echoed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: "Required management session or registered internal identity is missing, invalid or expired; surfaces are not interchangeable."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: "Authenticated actor lacks permission, recent authentication, a valid current-actor confirmation or required same-origin protection."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: "Business state, generation, idempotency or lease conflict, including COMPILE_OBSOLETE, IDEMPOTENCY_CONFLICT and LEASE_LOST. A stale resource revision is 412 instead."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    RateLimited:
      description: "Rate, slot, concurrent execution or reserved budget limit exceeded."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unavailable:
      description: "Primary database, execution capability or safe publication is unavailable. Fail closed; never serve cached authorization or substitute a direct/empty configuration."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    RunnerHeartbeatResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/RunnerHeartbeatResponse" }
...
    RunnerLeaseResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/RunnerLeaseResponse" }
...
  schemas:
    UUID:
      type: "string"
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    RequestID:
      type: "string"
      pattern: "^[A-Za-z0-9_-]{1,64}$"
    Revision:
      type: "string"
      pattern: "^(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)$"
      description: "Canonical positive decimal int64 string, 1..9223372036854775807. String encoding preserves exact values in browser clients. No signs or leading zeros."
...
    SchemaVersion:
      type: "integer"
      const: 1
    ShortText:
      type: "string"
      minLength: 1
      maxLength: 200
      pattern: "^[^\\x00-\\x1f\\x7f]+$"
...
    SHA256:
      type: "string"
      pattern: "^[0-9a-f]{64}$"
...
    Timestamp:
      type: "string"
      format: "date-time"
    SafeFieldPath:
      type: "string"
      maxLength: 512
      pattern: "^(?:/(?:[a-z][a-z0-9_]*|[0-9]+))*$"
      description: "Safe RFC 6901 pointer using only fixed, validated contract fields and array indices; root is empty. Unknown or duplicate input keys are represented by their safe parent, never echoed. Syntax alone is not authorization to output a supplied path."
    ErrorCode:
      type: "string"
      enum: ["MALFORMED_REQUEST", "UNKNOWN_FIELD", "DUPLICATE_FIELD", "AUTH_REQUIRED", "SESSION_EXPIRED", "PERMISSION_DENIED", "REAUTH_REQUIRED", "RESOURCE_NOT_FOUND", "STATE_CONFLICT", "IDEMPOTENCY_CONFLICT", "LEASE_LOST", "REVISION_MISMATCH", "INPUT_LIMIT_EXCEEDED", "VALIDATION_FAILED", "PRECONDITION_REQUIRED", "RATE_LIMITED", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR", "COMPILE_OBSOLETE", "CAPABILITY_UNSUPPORTED", "CAPABILITY_UNVERIFIED", "DEPENDENCY_EXCLUDED", "DEPENDENCY_CYCLE", "DNS_CYCLE", "EMPTY_GROUP", "PUBLICATION_BLOCKED", "PUBLICATION_CONFIRMATION_REQUIRED", "SUBSCRIPTION_NOT_READY", "SUBSCRIPTION_BLOCKED", "RUNNER_UNAVAILABLE", "BUDGET_EXCEEDED", "INVALID_CONFIG", "CORE_CONFIG_INVALID", "PROXY_CONNECT_FAILED", "AUTH_FAILED", "TARGET_TLS_FAILED", "HTTP_EXPECTATION_FAILED", "CORE_BINARY_MISMATCH", "PORT_BUSY", "RUNNER_RESOURCE_LIMIT", "CANCELED", "JOB_TIMEOUT", "TEST_TARGET_UNAVAILABLE", "INSUFFICIENT_SAMPLE"]
    ErrorDetail:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        field_path: { $ref: "#/components/schemas/SafeFieldPath" }
        resource_id: { $ref: "#/components/schemas/UUID" }
      minProperties: 1
    SafeError:
      type: "object"
      additionalProperties: false
      required: ["code", "message", "details"]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: "string"
          minLength: 1
          maxLength: 512
          description: "Bounded fixed safe explanation; no input values, SQL, ciphertext, URLs, raw schema messages or core stdout/stderr."
        details:
          type: "array"
          minItems: 0
          maxItems: 64
          items: { $ref: "#/components/schemas/ErrorDetail" }
    ErrorResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "error"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        error: { $ref: "#/components/schemas/SafeError" }
...
    CoreFamily:
      type: "string"
      enum: ["xray", "sing-box", "mihomo"]
...
    Platform:
      type: "string"
      enum: ["linux", "windows", "macos", "android", "ios"]
    Architecture:
      type: "string"
      enum: ["amd64", "arm64"]
    CoreIdentity:
      type: "object"
      additionalProperties: false
      required: ["core_build_id", "core_family", "version", "build_sha256", "platform", "architecture", "adapter_version"]
      properties:
        core_build_id: { $ref: "#/components/schemas/UUID" }
        core_family: { $ref: "#/components/schemas/CoreFamily" }
        version: { $ref: "#/components/schemas/ShortText" }
        build_sha256: { $ref: "#/components/schemas/SHA256" }
        platform: { $ref: "#/components/schemas/Platform" }
        architecture: { $ref: "#/components/schemas/Architecture" }
        adapter_version: { $ref: "#/components/schemas/ShortText" }
...
    RunnerJobType:
      type: "string"
      enum: ["config_validate", "connectivity", "download_throughput"]
...
    FrozenTestSubject:
      type: "object"
      additionalProperties: false
      required: ["kind", "id", "revision", "security_epoch"]
      properties:
        kind:
          type: "string"
          enum: ["node", "chain"]
        id: { $ref: "#/components/schemas/UUID" }
        revision: { $ref: "#/components/schemas/Revision" }
        security_epoch: { $ref: "#/components/schemas/Revision" }
...
    RunnerBuildReport:
      type: "object"
      additionalProperties: false
      required: ["core_build_id", "build_sha256"]
      properties:
        core_build_id: { $ref: "#/components/schemas/UUID" }
        build_sha256: { $ref: "#/components/schemas/SHA256" }
    RunnerSlots:
      type: "object"
      additionalProperties: false
      required: ["config_validate", "connectivity", "download_throughput"]
      properties:
        config_validate:
          type: "integer"
          minimum: 0
          maximum: 64
        connectivity:
          type: "integer"
          minimum: 0
          maximum: 64
        download_throughput:
          type: "integer"
          minimum: 0
          maximum: 16
    RunnerHeartbeatRequest:
      type: "object"
      additionalProperties: false
      required: ["runner_id", "platform", "architecture", "builds", "available_slots", "load"]
      properties:
        runner_id: { $ref: "#/components/schemas/UUID" }
        platform: { $ref: "#/components/schemas/Platform" }
        architecture: { $ref: "#/components/schemas/Architecture" }
        builds:
          type: "array"
          minItems: 0
          maxItems: 100
          items: { $ref: "#/components/schemas/RunnerBuildReport" }
        available_slots: { $ref: "#/components/schemas/RunnerSlots" }
        load:
          type: "object"
          additionalProperties: false
          required: ["active_jobs", "cpu_throttled", "memory_available_bytes"]
          properties:
            active_jobs:
              type: "integer"
              minimum: 0
              maximum: 144
            cpu_throttled:
              type: "boolean"
            memory_available_bytes:
              type: "integer"
              minimum: 0
              maximum: 9007199254740991
      description: "Certificate-bound identity and server registration are authoritative. No capabilities, location or binaries may be installed merely because this request advertises them."
    RunnerHeartbeatData:
      type: "object"
      additionalProperties: false
      required: ["runner_id", "heartbeat_interval_ms", "lease_duration_ms", "server_time", "accepted_build_ids"]
      properties:
        runner_id: { $ref: "#/components/schemas/UUID" }
        heartbeat_interval_ms:
          type: "integer"
          minimum: 1000
          maximum: 30000
        lease_duration_ms:
          type: "integer"
          minimum: 1000
          maximum: 120000
        server_time: { $ref: "#/components/schemas/Timestamp" }
        accepted_build_ids:
          type: "array"
          minItems: 0
          maxItems: 100
          items: { $ref: "#/components/schemas/UUID" }
          uniqueItems: true
    RunnerHeartbeatResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/RunnerHeartbeatData" }
    RunnerLeaseRequest:
      type: "object"
      additionalProperties: false
      required: ["runner_id", "available_slots"]
      properties:
        runner_id: { $ref: "#/components/schemas/UUID" }
        available_slots: { $ref: "#/components/schemas/RunnerSlots" }
      description: "Only capacity is requested. The server matches its registered certificate capabilities to queue types and exact builds. A client cannot nominate a job, URL, command, path or user configuration."
...
    RunnerLease:
      type: "object"
      additionalProperties: false
      required: ["job_id", "attempt", "lease_seq", "lease_expires_at", "schema_version", "type", "core", "artifact", "limits", "execution_policy", "payload_sha256"]
      properties:
        job_id: { $ref: "#/components/schemas/UUID" }
        attempt:
          type: "integer"
          minimum: 1
          maximum: 100
        lease_seq: { $ref: "#/components/schemas/Revision" }
        lease_expires_at: { $ref: "#/components/schemas/Timestamp" }
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        type: { $ref: "#/components/schemas/RunnerJobType" }
        core: { $ref: "#/components/schemas/CoreIdentity" }
        artifact: { $ref: "#/components/schemas/RunnerArtifact" }
        subject: { $ref: "#/components/schemas/FrozenTestSubject" }
        test_target: { $ref: "#/components/schemas/FrozenTestTarget" }
        limits: { $ref: "#/components/schemas/TestLimits" }
        execution_policy: { $ref: "#/components/schemas/RunnerExecutionPolicy" }
        payload_sha256: { $ref: "#/components/schemas/SHA256" }
        quota_reservation_id: { $ref: "#/components/schemas/UUID" }
      allOf:
        -
          if:
            properties:
              type:
                enum: ["connectivity", "download_throughput"]
          then:
            required: ["subject", "test_target", "quota_reservation_id"]
            properties:
              execution_policy:
                properties:
                  network:
                    const: "controlled_target_only"
        -
          if:
            properties:
              type:
                const: "config_validate"
          then:
            not:
              required: ["test_target"]
            properties:
              execution_policy:
                properties:
                  network:
                    const: "none"
      description: "Server-assigned bounded lease. payload_sha256 binds the canonical frozen payload. Paths, argv and controlled local listener ports are allocated by the registered Runner adapter, never supplied by users. Expired renewal requires local stop before expiry."
    RunnerLeaseData:
      type: "object"
      additionalProperties: false
      required: ["lease"]
      properties:
        lease:
          anyOf:
            - { $ref: "#/components/schemas/RunnerLease" }
            -
              type: "null"
    RunnerLeaseResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/RunnerLeaseData" }
...
paths:
...
  /internal/v1/runners/heartbeat:
    servers: [{url: 'https://runner-api.proxyloom.invalid', description: Dedicated internal mTLS listener; replace with the configured internal origin.}]
    post:
      operationId: heartbeatRunner
      x-implementation-status: implemented
      x-authentication-surface: runner-mtls
      security: [{RunnerMTLS: []}]
      summary: Report registered build digests and available slots from a certificate-bound Runner
      description: runner_id must match the registered client-certificate identity. Self-reported capabilities cannot expand the server registration.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunnerHeartbeatRequest'}
      responses:
        '200': {$ref: '#/components/responses/RunnerHeartbeatResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /internal/v1/jobs/lease:
    servers: [{url: 'https://runner-api.proxyloom.invalid', description: Dedicated internal mTLS listener.}]
    post:
      operationId: leaseRunnerJob
      x-implementation-status: implemented
      x-authentication-surface: runner-mtls
      security: [{RunnerMTLS: []}]
      summary: Let the server assign an eligible Runner task and frozen payload
      description: >-
        The request cannot choose job_id, arbitrary URL, shell, path or native
        configuration. Selection atomically checks certificate registration,
        exact core digest, type, slot and the attempt's budget reservation.
        Only config_validate, connectivity and download_throughput are eligible;
        source_refresh, import_parse and compile remain on the API Worker.
        No eligible task returns lease=null. Default lease is 30 seconds with
        5-second heartbeat; a Runner without a valid renewal stops before expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RunnerLeaseRequest'}
      responses:
        '200': {$ref: '#/components/responses/RunnerLeaseResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
...